Fixed test and fixtures.
