+++ composer.json
@@ -123,7 +123,7 @@
             "php-cs-fixer fix src"
         ],
         "analyse": [
-            "./vendor/bin/phpstan analyse --memory-limit 2G"
+            ".\\vendor\\bin\\phpstan analyse --memory-limit 2G"
         ]
     },
     "conflict": {
